--- plugins/sonic-go-mitmproxy-ca-cert.pem
@@ -1,21 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDdDCCAlygAwIBAgIGDvq8Ws58MA0GCSqGSIb3DQEBCwUAMCgxEjAQBgNVBAoT
-CW1pdG1wcm94eTESMBAGA1UEAxMJbWl0bXByb3h5MB4XDTIyMDMwOTEzNDc0OFoX
-DTI1MDMxMDEzNDc0OFowKDESMBAGA1UEChMJbWl0bXByb3h5MRIwEAYDVQQDEwlt
-aXRtcHJveHkwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCz6fzRW1iU
-Y1vzY0Uzmi0KWMi4OhK1iiRT/Km+Imr9gTdFqTT8MixcnGzKHtFa+UONEZdWt8cT
-DCMa4jeUzT+Xivl6gxbuMFKh5qCx5LrJqLGTTf4ArcARGhEWoFJuqk3fCagNWtaC
-5VQU+xmRubDK7vbvzYQXKhuP89XUznyB8k0hEP5MOquExGVvlYR1PeOgRt/PYLGF
-XUyhZtxCrTvtmfESfR6QrAF075ikEHfGpHAEYLSVizGftFE36s0ZJBom0eC0mxUL
-jat7Ml6qzjcrPIiJtz96woCfbR31Ifi+G03XqpcdRm3yJdCB21PSNOdddCzk223y
-urkjJHQIoTX/AgMBAAGjgaMwgaAwDgYDVR0PAQH/BAQDAgEGMF4GA1UdJQRXMFUG
-CCsGAQUFBwMBBggrBgEFBQcDAgYIKwYBBQUHAwQGCCsGAQUFBwMIBggrBgEFBQcD
-AwYKKwYBBAGCNwIBFgYKKwYBBAGCNwoDAwYJYIZIAYb4QgQBMA8GA1UdEwEB/wQF
-MAMBAf8wHQYDVR0OBBYEFCL01AYSxccm0gW1GL5bgHjeh/7aMA0GCSqGSIb3DQEB
-CwUAA4IBAQAJni+7Vr5sXSUJoFD06cklGjwdxZ81VSxMoieIOo7Pni3l+I/HDDRQ
-MhlH/+lkMPfrGmHHV29BPQpaChlYPkToLU0ZmbxgWy45mzV5/0fzMV8/1VsM5S2Q
-ZtUFRp9gpWIjJcsUYvtrroHh8WQmF8gcikU99KiTEJ9mhdj6IMtlXYgtXiaTnm+L
-XGslcizMfQy2p5bk+PzYNsyWYYZ+ftb0dGXKAT2oGu4Xd+SHJVTQ5CQwpj/Cr01H
-rUkmnazqxYiiNRswq9qpa0+dpCURqA40K0I4R/4Q1W14KVA/2KMnUPPISMtpdpA6
-hz0dycUmiiqsT2zR1iTkaIxpVHcHns08
------END CERTIFICATE-----
