--- plugins/sonic-go-mitmproxy-ca.pem
@@ -1,49 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCz6fzRW1iUY1vz
-Y0Uzmi0KWMi4OhK1iiRT/Km+Imr9gTdFqTT8MixcnGzKHtFa+UONEZdWt8cTDCMa
-4jeUzT+Xivl6gxbuMFKh5qCx5LrJqLGTTf4ArcARGhEWoFJuqk3fCagNWtaC5VQU
-+xmRubDK7vbvzYQXKhuP89XUznyB8k0hEP5MOquExGVvlYR1PeOgRt/PYLGFXUyh
-ZtxCrTvtmfESfR6QrAF075ikEHfGpHAEYLSVizGftFE36s0ZJBom0eC0mxULjat7
-Ml6qzjcrPIiJtz96woCfbR31Ifi+G03XqpcdRm3yJdCB21PSNOdddCzk223yurkj
-JHQIoTX/AgMBAAECggEAYsX3csj/nJfgB4NcnrPqPl/38oyPcPNidmmklyo2Ku8h
-9TCsUjkJed63K1A4CAYJPA2IVKeYt7BWJZBEi4EThJNb3BLUJ/iHpTMEziWRFt1y
-gDi1VpINfFYfQ5jJKNB+6BBZHtvcIQjeXJZutz6G/SKsl4LxHfaUIzfvEWI7oOsG
-A86shQrpRG/f/VcNPV5NCjwD1F2iWi8sDprWIWxQ5n8GKT1a9V1VqSGR/LEEvQmK
-kjwe9VjZqCgVYyil0F2yulEVbrRvAjCIu30SfYgMBJ2s6WtrV9OLyCJkLn7RlHLQ
-+Ht5n86fQ+iFRbs39GPlIHZSEOHwLf3i/pDCbMrEQQKBgQDQOOhRLaj4LEQLNK6F
-/vWWAF6j+tF0ZWrFmtEOiDokGl9o7Dddjh21SCYlRb21Z5eeB21aNNQTV4iUZ/Tt
-A2x1VMT5YQ58htovaWGWZcx2peTTBerq6j0fE1T6EOykhClf9ph4qR6P2/JXBmWi
-PD1s+C1RvLUt9CkwYqStCiN7BwKBgQDdMjxs4vo3+aKdljz0PefrbtWMkcC9gqL+
-WNCRnlA9vhyWQ4+/y5jVtLCrBUc5Aelya45z8/+dF/mqO7JNw3E2DZcxm/BWNzxM
-/wTZucLfT2X8N2/bh0eC4GDMB0AARhhBTRpM7zdT+sRULu02+1fcuScwtsrMs9Ex
-WOfSKv6XSQKBgGMKvo5mfpnVTGRcIauA88Xu1RWAp4KK/veRRWF5MNadaMaanVc1
-0zG4aNSP0MXpyXeUptIikMHGgCO/MwRN/Zftw2evX9mht0SYBqAoOdIx5J7vkKc+
-1eXjJydvR3CVoKyRukKihvP8tc/pLgy2RK2BeuUHMtLx8Wmfch8Q1IGfAoGBALYJ
-tN7rwX81i9XjIdDj7ay9LzQbc6LDEz79etCMmH7nqYEkovX2KZaeuNJWarII18WZ
-rMfSo6f2ZjrqHCiRUudoMeja6+UYWiEHs0c38J91KXpzaTQi1gBohmEDCnHvPeDo
-AmAUctVDSDX71K9uFHYLTqxxkt76kYvqHgbB6Xo5AoGBAJ2g4kWINE6Mw577I0Sa
-ODEgnPQdNK5NaCboUOqkzdIw0Sll7lyXXlibladH5J4SSMZaiJ/QvEZrZKT6ejkD
-5+yh+TuRqKImiJRoWZ3r/9pxJyveMuaSD9w3rYkeV4Eq+CnRX2QU+azeqVRznSsC
-bUffrBRDx4QvYII/owrQjehw
------END PRIVATE KEY-----
------BEGIN CERTIFICATE-----
-MIIDdDCCAlygAwIBAgIGDvq8Ws58MA0GCSqGSIb3DQEBCwUAMCgxEjAQBgNVBAoT
-CW1pdG1wcm94eTESMBAGA1UEAxMJbWl0bXByb3h5MB4XDTIyMDMwOTEzNDc0OFoX
-DTI1MDMxMDEzNDc0OFowKDESMBAGA1UEChMJbWl0bXByb3h5MRIwEAYDVQQDEwlt
-aXRtcHJveHkwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQCz6fzRW1iU
-Y1vzY0Uzmi0KWMi4OhK1iiRT/Km+Imr9gTdFqTT8MixcnGzKHtFa+UONEZdWt8cT
-DCMa4jeUzT+Xivl6gxbuMFKh5qCx5LrJqLGTTf4ArcARGhEWoFJuqk3fCagNWtaC
-5VQU+xmRubDK7vbvzYQXKhuP89XUznyB8k0hEP5MOquExGVvlYR1PeOgRt/PYLGF
-XUyhZtxCrTvtmfESfR6QrAF075ikEHfGpHAEYLSVizGftFE36s0ZJBom0eC0mxUL
-jat7Ml6qzjcrPIiJtz96woCfbR31Ifi+G03XqpcdRm3yJdCB21PSNOdddCzk223y
-urkjJHQIoTX/AgMBAAGjgaMwgaAwDgYDVR0PAQH/BAQDAgEGMF4GA1UdJQRXMFUG
-CCsGAQUFBwMBBggrBgEFBQcDAgYIKwYBBQUHAwQGCCsGAQUFBwMIBggrBgEFBQcD
-AwYKKwYBBAGCNwIBFgYKKwYBBAGCNwoDAwYJYIZIAYb4QgQBMA8GA1UdEwEB/wQF
-MAMBAf8wHQYDVR0OBBYEFCL01AYSxccm0gW1GL5bgHjeh/7aMA0GCSqGSIb3DQEB
-CwUAA4IBAQAJni+7Vr5sXSUJoFD06cklGjwdxZ81VSxMoieIOo7Pni3l+I/HDDRQ
-MhlH/+lkMPfrGmHHV29BPQpaChlYPkToLU0ZmbxgWy45mzV5/0fzMV8/1VsM5S2Q
-ZtUFRp9gpWIjJcsUYvtrroHh8WQmF8gcikU99KiTEJ9mhdj6IMtlXYgtXiaTnm+L
-XGslcizMfQy2p5bk+PzYNsyWYYZ+ftb0dGXKAT2oGu4Xd+SHJVTQ5CQwpj/Cr01H
-rUkmnazqxYiiNRswq9qpa0+dpCURqA40K0I4R/4Q1W14KVA/2KMnUPPISMtpdpA6
-hz0dycUmiiqsT2zR1iTkaIxpVHcHns08
------END CERTIFICATE-----
