+++ src/main/java/org/cloud/sonic/agent/tests/IOSTests.java
@@ -3,7 +3,7 @@ package org.cloud.sonic.agent.tests;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
 import org.cloud.sonic.agent.automation.IOSStepHandler;
-import org.cloud.sonic.agent.bridge.ios.TIDeviceTool;
+import org.cloud.sonic.agent.bridge.ios.SibTool;
 import org.cloud.sonic.agent.common.interfaces.DeviceStatus;
 import org.cloud.sonic.agent.tests.ios.IOSTestTaskBootThread;
 import org.slf4j.Logger;
@@ -35,7 +35,7 @@ public class IOSTests {
         List<JSONObject> dataProvider = new ArrayList<>();
         for (JSONObject device : dataInfo.getJSONArray("device").toJavaList(JSONObject.class)) {
             String udId = device.getString("udId");
-            if (!TIDeviceTool.getDeviceList().contains(udId)) {
+            if (!SibTool.getDeviceList().contains(udId)) {
                 continue;
             }
             JSONObject deviceTestData = new JSONObject();
